extension(delegate) 코드 분리
